requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-groq>=0.0.1
PyPDF2==3.0.1
faiss-cpu==1.9.0
langchain_google_genai
scikit-learn
streamlit-option-menu
llama-index
pathway
docutils==0.21.2
docx2txt==0.8
duckduckgo_search==6.3.3
email_validator==2.2.0
huggingface-hub==0.25.1
langchain==0.3.4
langchain-huggingface==0.1.0
langchain-ollama==0.2.0
langchain-openai==0.2.3
langgraph==0.2.43
nltk==3.9.1
ollama==0.3.3
openai==1.52.0
PyMuPDF==1.24.12
PyMySQL==1.1.1
tavily-python==0.5.0
textblob
wikipedia==1.4.0
youtube-transcript-api==0.6.2
SpeechRecognition>=3.10.0
gTTS>=2.5.0
pygame>=2.5.2
pyttsx3>=2.90
# pip install streamlit pathway[all] litellm==1.40.0 google-generativeai
pathway
streamlit>=1.30.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-groq>=0.0.1
PyPDF2==3.0.1
faiss-cpu==1.9.0
langchain_google_genai
scikit-learn
streamlit-option-menu
llama-index
pathway
docutils==0.21.2
docx2txt==0.8
duckduckgo_search==6.3.3
email_validator==2.2.0
huggingface-hub==0.25.1
langchain==0.3.4
langchain-huggingface==0.1.0
langchain-ollama==0.2.0
langchain-openai==0.2.3
langgraph==0.2.43
nltk==3.9.1
ollama==0.3.3
openai==1.52.0
PyMuPDF==1.24.12
PyMySQL==1.1.1
tavily-python==0.5.0
textblob
wikipedia==1.4.0
youtube-transcript-api==0.6.2
SpeechRecognition>=3.10.0
gTTS>=2.5.0
pygame>=2.5.2
pyttsx3>=2.90
# pip install streamlit pathway[all] litellm==1.40.0 google-generativeai
